refactoring the habits
