+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/habit/database/HabitDao.kt
@@ -6,7 +6,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.example.habittrackerapp.model.Habit
+import com.example.habittrackerapp.model.habitViewModel.Habit
 import kotlinx.coroutines.flow.Flow
 
 @Dao
