+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/habit/database/HabitDatabase.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import com.example.habittrackerapp.model.Habit
+import com.example.habittrackerapp.model.habitViewModel.Habit
 
 @Database(entities = [Habit::class], version = 1, exportSchema = false)
 abstract class HabitDatabase : RoomDatabase() {
