+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/habit/screens/HabitsForTodayListScreen.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.habittrackerapp.LocalNavController
-import com.example.habittrackerapp.model.Habit
+import com.example.habittrackerapp.model.habitViewModel.Habit
 import com.example.habittrackerapp.model.habitViewModel.HabitCreateAndListViewModel
 import com.example.habittrackerapp.model.habitViewModel.HabitViewModelProvider
