+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/habit/screens/SaveChanges.kt
@@ -8,11 +8,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.habittrackerapp.LocalNavController
-import com.example.habittrackerapp.model.Habit
-import com.example.habittrackerapp.model.habitViewModel.HabitCreateAndListViewModel
 import com.example.habittrackerapp.model.habitViewModel.HabitUpdateViewModel
 import com.example.habittrackerapp.model.habitViewModel.HabitViewModelProvider
-import com.example.habittrackerapp.model.habitViewModel.toHabitDetails
 
 
 @Composable
