+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/habitViewModel/HabitCreateAndListViewModel.kt
@@ -5,8 +5,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.habittrackerapp.model.Habit
-import com.example.habittrackerapp.model.HabitRepository
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
