+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/habitViewModel/HabitReadAndDeleteViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.habittrackerapp.habit.Frequency
-import com.example.habittrackerapp.model.HabitRepository
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.filterNotNull
