+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/habitViewModel/HabitUpdateViewModel.kt
@@ -6,13 +6,8 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.habittrackerapp.model.HabitRepository
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 
 class HabitUpdateViewModel (savedStateHandle: SavedStateHandle,
