+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/userViewModel/SavedUser.kt
@@ -1,4 +1,4 @@
-package com.example.habittrackerapp.model
+package com.example.habittrackerapp.model.userViewModel
 
 data class SavedUser (
     var email: String = "",
