+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/userViewModel/SavedUserRepository.kt
@@ -1,6 +1,5 @@
 package com.example.habittrackerapp.model.userViewModel
 
-import com.example.habittrackerapp.model.SavedUser
 import kotlinx.coroutines.flow.Flow
 
 interface SavedUserRepository {
