+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/userViewModel/SavedUserViewModel.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.example.habittrackerapp.MyApp
-import com.example.habittrackerapp.model.SavedUser
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
