+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/userViewModel/UserRepositoryDatastore.kt
@@ -6,7 +6,6 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
-import com.example.habittrackerapp.model.SavedUser
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
